Fixed a potential error
If we log in successfully, the LoginPage is unmounted and the ProfilePage is mounted. The chakra.form has a onSubmit function which ends with a finally clause, leading the setInSuccessfully state to change which will lead to errors.

We created useMounted hook for the purpose of cheking if the component is still mounted or not.